did not commit long time)

+++ NBFiLib/nbfi/nbfi_misc.c
@@ -157,7 +157,7 @@ void NBFi_RxPacket_Free(nbfi_transport_packet_t* pkt)
 
 uint8_t NBFi_Packets_To_Send()
 {
-
+    //return 0;
     uint8_t packets_free = 0;
 
     for(uint16_t i = nbfi_TXbuf_head; i != (nbfi_TXbuf_head + NBFI_TX_PKTBUF_SIZE); i++)
@@ -512,13 +512,13 @@ uint8_t CRC8(uint8_t* data, uint8_t len)
 #define POLYNOMIAL (0x104C11DB7)
 
 uint32_t crc_table(uint8_t n)
-{
+{ 
     uint32_t c;
     int k;
     c=((uint32_t)n) << (WIDTH - 8);
     for(k=8;k>0;k--)
     {
-        if(c & TOPBIT)
+             if(c & (uint32_t)TOPBIT)
         {
             c = (c<<1) ^ POLYNOMIAL;
         }
